Refactor clam_adapter to use SNSWriter instead of SNSMessage. Deleted SNSMessage and PublishableMessage
