--- common/src/main/scala/uk/ac/wellcome/models/PublishableMessage.scala
@@ -1,9 +1,0 @@
-package uk.ac.wellcome.models
-
-import scala.util.Try
-
-case class PublishAttempt(id: String)
-
-trait PublishableMessage extends Message {
-  def publish(): Try[PublishAttempt]
-}
