+++ common/src/main/scala/uk/ac/wellcome/sns/SNSWriter.scala
@@ -4,12 +4,12 @@ import com.amazonaws.services.sns.AmazonSNS
 import com.amazonaws.services.sns.model.PublishRequest
 import com.google.inject.Inject
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.PublishAttempt
 import uk.ac.wellcome.models.aws.SNSConfig
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
+case class PublishAttempt(id: String)
 
 class SNSWriter @Inject()(snsClient: AmazonSNS, snsConfig: SNSConfig) extends Logging {
   val defaultSubject = "subject-not-specified"
@@ -17,6 +17,11 @@ class SNSWriter @Inject()(snsClient: AmazonSNS, snsConfig: SNSConfig) extends Lo
   def writeMessage(message: String, subject: Option[String]): Future[PublishAttempt] = Future {
     info(s"about to publish message $message on the SNS topic ${snsConfig.topicArn}")
     snsClient.publish(new PublishRequest(snsConfig.topicArn, message, subject.getOrElse(defaultSubject)))}.map{publishResult =>
+      info(s"Published message ${publishResult.getMessageId}")
       PublishAttempt(publishResult.getMessageId)
-    }
+    }.recover{
+    case e:Throwable=>
+      error("Failed to publish message", e)
+      throw e
+  }
 }
