+++ id_minter/TODO
@@ -1,4 +1,3 @@
-- Refactor transformer with SNSWriter
 - Remove id from UnifiedItem
 - create test_common module
 - remove global execution context
