+++ id_minter/src/test/scala/uk/ac/wellcome/platform/idminter/steps/IdGeneratorTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.idminter.modules
+package uk.ac.wellcome.platform.idminter.steps
 
 import com.gu.scanamo.Scanamo
 import com.gu.scanamo.syntax._
@@ -6,7 +6,6 @@ import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
 import uk.ac.wellcome.models.{Id, Identifier, UnifiedItem}
 import uk.ac.wellcome.platform.idminter.DynamoDBLocal
-import uk.ac.wellcome.platform.idminter.steps.IdGenerator
 
 class IdGeneratorTest extends FunSpec with DynamoDBLocal with ScalaFutures with Matchers with BeforeAndAfterEach with IntegrationPatience{
