+++ transformer/src/main/scala/uk/ac/wellcome/transformer/receive/RecordReceiver.scala
@@ -2,30 +2,27 @@ package uk.ac.wellcome.transformer.receive
 
 import com.amazonaws.services.dynamodbv2.model.AttributeValue
 import com.amazonaws.services.dynamodbv2.streamsadapter.model.RecordAdapter
-import com.amazonaws.services.sns.AmazonSNS
 import com.google.inject.Inject
 import com.gu.scanamo.ScanamoFree
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.aws.{SNSConfig, SNSMessage}
 import uk.ac.wellcome.models.{CalmDynamoRecord, Transformable, UnifiedItem}
+import uk.ac.wellcome.sns.{PublishAttempt, SNSWriter}
 import uk.ac.wellcome.utils.JsonUtil
 
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
-import scala.concurrent.ExecutionContext.Implicits.global
 
 case class RecordMap(value: java.util.Map[String, AttributeValue])
 
-class RecordReceiver @Inject()(snsConfig: SNSConfig,
-                               snsClient: AmazonSNS) extends Logging {
+class RecordReceiver @Inject()(snsWriter: SNSWriter) extends Logging {
 
   def receiveRecord(record: RecordAdapter): Future[Unit] = {
     for {
       recordMap           <- recordToRecordMap(record)
       transformableRecord <- extractTransformableCaseClass(recordMap)
       cleanRecord         <- transformDynamoRecord(transformableRecord)
-      snsMessage          <- buildSNSMessageFrom(cleanRecord)
-      _                   <- publishMessage(snsMessage)
+      _                   <- publishMessage(cleanRecord)
     } yield ()
   }
 
@@ -62,30 +59,6 @@ class RecordReceiver @Inject()(snsConfig: SNSConfig,
     }
   }
 
-  def buildSNSMessageFrom(unifiedItem: UnifiedItem): Future[SNSMessage] = {
-    Future{JsonUtil.toJson(unifiedItem)}.map {
-      case Success(stringifiedJson) =>
-        val message = SNSMessage(
-          Some("Foo"),
-          stringifiedJson,
-          snsConfig.topicArn,
-          snsClient
-        )
-        info(s"Publishable message $message")
-        message
-      case Failure(e) =>
-        error("Failed to convert into publishable message", e)
-        throw e
-    }
-  }
-
-  def publishMessage(message: SNSMessage): Future[Unit] = Future{
-    message.publish() match {
-      case Success(publishAttempt) =>
-        info(s"Published message ${publishAttempt.id}")
-      case Failure(e) =>
-        error("Failed to publish message", e)
-        throw e
-    }
-  }
+  def publishMessage(unifiedItem: UnifiedItem): Future[PublishAttempt] =
+    snsWriter.writeMessage(JsonUtil.toJson(unifiedItem).get, Some("Foo"))
 }
